requirements.txt
Django>=5.0,<6.0
django-ninja>=1.1.0
aiogram>=3.12
pydantic>=2.7
httpx>=0.27
celery>=5.3
redis>=5.0
sentry-sdk>=2.0
psycopg2-binary>=2.9
dj-database-url>=2.2
supabase>=2.5
structlog>=24.1
uvicorn[standard]>=0.30
gunicorn>=21.2
orjson>=3.10
python-dotenv>=1.0
flower>=2.0